--- a/src/Microsoft.AspNetCore.Routing/Matching/CandidateSet.cs
+++ b/src/Microsoft.AspNetCore.Routing/Matching/CandidateSet.cs
@@ -2,6 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Collections;
+using System.Collections.Specialized;
 using System.Runtime.CompilerServices;
 using Microsoft.AspNetCore.Http;
 
@@ -14,6 +16,11 @@ namespace Microsoft.AspNetCore.Routing.Matching
     /// </summary>
     public sealed class CandidateSet
     {
+        private const int BitVectorSize = 32;
+
+        private BitVector32 _validity;
+        private BitArray _largeCapactityValidity;
+
         // We inline storage for 4 candidates here to avoid allocations in common
         // cases. There's no real reason why 4 is important, it just seemed like 
         // a plausible number.
@@ -26,8 +33,7 @@ namespace Microsoft.AspNetCore.Routing.Matching
 
         /// <summary>
         /// <para>
-        /// Initializes a new instances of the candidate set structure with the provided list of endpoints 
-        /// and associated scores.
+        /// Initializes a new instances of the candidate set structure with the provided data.
         /// </para>
         /// <para>
         /// The constructor is provided to enable unit tests of implementations of <see cref="EndpointSelector"/>
@@ -35,8 +41,9 @@ namespace Microsoft.AspNetCore.Routing.Matching
         /// </para>
         /// </summary>
         /// <param name="endpoints">The list of endpoints, sorted in descending priority order.</param>
+        /// <param name="values">The list of <see cref="RouteValueDictionary"/> instances.</param>
         /// <param name="scores">The list of endpoint scores. <see cref="CandidateState.Score"/>.</param>
-        public CandidateSet(Endpoint[] endpoints, int[] scores)
+        public CandidateSet(Endpoint[] endpoints, RouteValueDictionary[] values, int[] scores)
         {
             Count = endpoints.Length;
 
@@ -79,6 +86,17 @@ namespace Microsoft.AspNetCore.Routing.Matching
                         _additionalCandidates[i - 4] = new CandidateState(endpoints[i], score: scores[i]);
                     }
                     break;
+            }
+
+            // Initialize validity to valid by default.
+            if (Count < BitVectorSize)
+            {
+                // Sets the bit for each candidate that exists (bits > Count will be 0).
+                _validity = new BitVector32(unchecked((int)~(0xFFFFFFFFu << Count)));
+            }
+            else
+            {
+                _largeCapactityValidity = new BitArray(Count, defaultValue: true);
             }
         }
 
@@ -126,25 +144,35 @@ namespace Microsoft.AspNetCore.Routing.Matching
                     }
                     break;
             }
+
+            // Initialize validity to valid by default.
+            if (Count < BitVectorSize)
+            {
+                // Sets the bit for each candidate that exists (bits > Count will be 0).
+                _validity = new BitVector32(unchecked((int)~(0xFFFFFFFFu << Count)));
+            }
+            else
+            {
+                _largeCapactityValidity = new BitArray(Count, defaultValue: true);
+            }
         }
 
         /// <summary>
         /// Gets the count of candidates in the set.
         /// </summary>
         public int Count { get; }
-
+        
         /// <summary>
         /// Gets the <see cref="CandidateState"/> associated with the candidate <see cref="Endpoint"/>
         /// at <paramref name="index"/>.
         /// </summary>
         /// <param name="index">The candidate index.</param>
         /// <returns>
-        /// A reference to the <see cref="CandidateState"/>. The result is returned by reference
-        /// and intended to be mutated.
+        /// A reference to the <see cref="CandidateState"/>. The result is returned by reference.
         /// </returns>
         public ref CandidateState this[int index]
         {
-            // Note that this is a ref-return because of both mutability and performance.
+            // Note that this is a ref-return because of performance.
             // We don't want to copy these fat structs if it can be avoided.
 
             // PERF: Force inlining
@@ -174,6 +202,56 @@ namespace Microsoft.AspNetCore.Routing.Matching
                     default:
                         return ref _additionalCandidates[index - 4];
                 }
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets a value which indicates where the <see cref="Http.Endpoint"/> is considered
+        /// a valid candiate for the current request. Set this value to <c>false</c> to exclude an
+        /// <see cref="Http.Endpoint"/> from consideration.
+        /// </summary>
+        public bool IsValidCandidate(int index)
+        {
+            // Friendliness for inlining
+            if ((uint)index >= Count)
+            {
+                ThrowIndexArgumentOutOfRangeException();
+            }
+
+            if (Count < BitVectorSize)
+            {
+                // Get the n-th bit
+                return _validity[0x00000001 << index];
+            }
+            else
+            {
+                return _largeCapactityValidity[index];
+            }
+        }
+
+        /// <summary>
+        /// Sets the validitity of the candidate at the provided index.
+        /// </summary>
+        /// <param name="index">The candidate index.</param>
+        /// <param name="value">
+        /// The value to set. If <c>true</c> the candidate is considered valid for the current request.
+        /// </param>
+        public void SetValidity(int index, bool value)
+        {
+            // Friendliness for inlining
+            if ((uint)index >= Count)
+            {
+                ThrowIndexArgumentOutOfRangeException();
+            }
+
+            if (Count < BitVectorSize)
+            {
+                // Set the n-th bit
+                _validity[0x00000001 << index] = value;
+            }
+            else
+            {
+                _largeCapactityValidity[index] = value;
             }
         }