… benchmarks/Microsoft.AspNetCore.Routing.Performance/EndpointRoutingBenchmarkBase.cs
… benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherAzureBenchmark.cs
… benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherFindCandidateSetSmallEntryCountBenchmark.cs
… benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherGithubBenchmark.cs
… benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherSingleEntryBenchmark.cs
… benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/TrivialMatcher.cs
… src/Microsoft.AspNetCore.Routing/Matching/EndpointSelector.cs
… src/Microsoft.AspNetCore.Routing/Matching/IEndpointSelectorPolicy.cs
… src/Microsoft.AspNetCore.Routing/Matching/Matcher.cs
… test/Microsoft.AspNetCore.Routing.Tests/EndpointMiddlewareTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/EndpointRoutingMiddlewareTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/EndpointSelectorContextTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/LinkGeneratorEndpointNameExtensionsTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/LinkGeneratorRouteValuesAddressExtensionsTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/LinkGeneratorTestBase.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/BarebonesMatcher.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/CandidateSetTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/DataSourceDependentMatcherTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/DefaultEndpointSelectorTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/DfaMatcherConformanceTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/DfaMatcherTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/FullFeaturedMatcherConformanceTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/HttpMethodMatcherPolicyIntegrationTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/MatcherAssert.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/MatcherConformanceTest.SingleEndpoint.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/MatcherConformanceTest.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/RouteMatcher.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/RouteMatcherBuilder.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/TreeRouterMatcher.cs
… test/Microsoft.AspNetCore.Routing.Tests/Matching/TreeRouterMatcherBuilder.cs