@rynowak 2018-09-30
https://github.com/aspnet/Mvc/pull/8522